:sparkles: Add support for multi-rows YAML
In order to make things a bit easier to handle, I
started splitting out the commands into individual files
