+++ misc/test-data/1.yaml
@@ -1,0 +1,10 @@
+a: 1
+b: 2
+c:
+  - 3
+  - 4
+  - 5
+d:
+  e: 6
+  f: 7
+
