+++ misc/test-data/2.yaml
@@ -1,0 +1,10 @@
+- a: 10
+  b: 20
+  c:
+    - 30
+    - 40
+    - 50
+  d:
+    e: 60
+    f: 70
+
