+++ misc/test-data/3.yaml
@@ -1,0 +1,5 @@
+- a: 100
+  b: 200
+  c:
+    - 300
+
