+++ misc/test-data/rows.json
@@ -1,0 +1,35 @@
+[
+  {
+    "a": 1,
+    "b": 2,
+    "c": [
+      3,
+      4,
+      5
+    ],
+    "d": {
+      "e": 6,
+      "f": 7
+    }
+  },
+  {
+    "a": 10,
+    "b": 20,
+    "c": [
+      30,
+      40,
+      50
+    ],
+    "d": {
+      "e": 60,
+      "f": 70
+    }
+  },
+  {
+    "a": 100,
+    "b": 200,
+    "c": [
+      300
+    ]
+  }
+]
